# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AMS - Advanced Medical Stomatology System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: patientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Delete appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # TREATMENTS
...
  /imaging:
    get:
      operationId: listImages
      tags: [imaging]
      summary: List medical images
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MedicalImage"
    post:
      operationId: createImage
      tags: [imaging]
      summary: Add medical image record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicalImageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalImage"
  /imaging/{id}:
    delete:
      operationId: deleteImage
      tags: [imaging]
      summary: Delete image
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # DENTAL CHART
  /dental-chart/{patientId}:
    get:
      operationId: getDentalChart
      tags: [dental-chart]
      summary: Get dental chart for patient
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dental chart
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DentalChartEntry"
  /dental-chart/{patientId}/tooth/{toothNumber}:
    put:
      operationId: updateToothStatus
      tags: [dental-chart]
      summary: Update tooth status
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
        - name: toothNumber
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToothStatusInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DentalChartEntry"
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [Administrator, Doctor, Receptionist, Assistant, Accountant, Manager]
        avatar:
          type: ["string", "null"]
...
    Appointment:
      type: object
      required: [id, patientId, patientName, date, time, duration, type, status]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: string
        patientAvatar:
          type: ["string", "null"]
        date:
          type: string
        time:
          type: string
        duration:
          type: integer
        type:
          type: string
          enum: [checkup, cleaning, filling, extraction, rootCanal, crown, implant, orthodontics, emergency, consultation, other]
        status:
          type: string
          enum: [scheduled, confirmed, in_progress, completed, cancelled, no_show]
        notes:
          type: ["string", "null"]
        doctorId:
          type: ["integer", "null"]
        doctorName:
          type: ["string", "null"]
    AppointmentInput:
      type: object
      required: [patientId, date, time, duration, type]
      properties:
        patientId:
          type: integer
        date:
          type: string
        time:
          type: string
        duration:
          type: integer
        type:
          type: string
          enum: [checkup, cleaning, filling, extraction, rootCanal, crown, implant, orthodontics, emergency, consultation, other]
        notes:
          type: string
        doctorId:
          type: integer
    AppointmentUpdate:
      type: object
      properties:
        date:
          type: string
        time:
          type: string
        duration:
          type: integer
        type:
          type: string
          enum: [checkup, cleaning, filling, extraction, rootCanal, crown, implant, orthodontics, emergency, consultation, other]
        status:
          type: string
          enum: [scheduled, confirmed, in_progress, completed, cancelled, no_show]
        notes:
          type: string
        doctorId:
          type: integer
...
    MedicalImage:
      type: object
      required: [id, patientId, type, url, date, createdAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        type:
          type: string
          enum: [panoramic, periapical, bitewing, cephalometric, cbct, intraoral, other]
        url:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        date:
          type: string
        description:
          type: ["string", "null"]
        toothNumber:
          type: ["integer", "null"]
        createdAt:
          type: string
    MedicalImageInput:
      type: object
      required: [patientId, type, url, date]
      properties:
        patientId:
          type: integer
        type:
          type: string
          enum: [panoramic, periapical, bitewing, cephalometric, cbct, intraoral, other]
        url:
          type: string
        thumbnailUrl:
          type: string
        date:
          type: string
        description:
          type: string
        toothNumber:
          type: integer
    DentalChartEntry:
      type: object
      required: [id, patientId, toothNumber, status, surface]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        toothNumber:
          type: integer
        status:
          type: string
          enum: [healthy, decayed, filled, crowned, missing, implant, extracted, root_canal, bridge, fractured]
        surface:
          type: string
        notes:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        updatedAt:
          type: string
    ToothStatusInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [healthy, decayed, filled, crowned, missing, implant, extracted, root_canal, bridge, fractured]
        surface:
          type: string
        notes:
          type: string
        color:
          type: string