# lib/api-zod/src/generated/types/dentalChartEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AMS - Advanced Medical Stomatology System API
 * OpenAPI spec version: 0.1.0
 */
import type { DentalChartEntryStatus } from './dentalChartEntryStatus';

export interface DentalChartEntry {
  id: number;
  patientId: number;
  toothNumber: number;
  status: DentalChartEntryStatus;
  surface: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  color?: string | null;
  updatedAt?: string;
}